+++ lib/sec_id/sedol.rb
@@ -45,8 +45,8 @@ module SecId
         raise InvalidFormatError, "'#{country_code}' is not a valid SEDOL country code!"
       end
 
-      restore!
-      isin = ISIN.new("#{country_code}00#{full_number}")
+      sedol_with_check_digit = "#{identifier}#{check_digit || calculate_check_digit}"
+      isin = ISIN.new("#{country_code}00#{sedol_with_check_digit}")
       isin.restore!
       isin
     end