+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- `CUSIP#to_isin` and `SEDOL#to_isin` no longer mutate source instance when check digit is missing ([#127](https://github.com/svyatov/sec_id/issues/127))
+
 ## [4.4.0] - 2026-01-29
 
 ### Added